+++ res/shaders/batch/vert.glsl
@@ -6,11 +6,11 @@ layout(location = 0) uniform mat4 u_proj;
 layout(location = 0) in vec2 v_pos;
 layout(location = 1) in vec4 v_color;
 layout(location = 2) in vec2 v_uv;
-layout(location = 3) in float v_tex;
+layout(location = 3) in int v_tex;
 
 out vec4 pass_color;
 out vec2 pass_uv;
-flat out float pass_tex;
+flat out int pass_tex;
 
 void main() {
     pass_color = v_color;
